Uses gpio_handler to refactor the user_button driver.

+++ drivers/inc/user_button.h
@@ -1,9 +1,9 @@
 #ifndef USER_BUTTON_H_
 #define USER_BUTTON_H_
 
-#include "stm32f4xx.h"
+#include <stdint.h>
 
 void user_button_init(void);
-int user_button_get_state(void);
+uint8_t user_button_get_state(void);
 
 #endif /* USER_BUTTON_H_ */
